% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LEGIT.R
\docType{data}
\name{best_model}
\alias{best_model}
\title{Best model}
\format{An object of class \code{function} of length 1.}
\usage{
best_model
}
\arguments{
\item{object}{An object}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Best model
}
\description{
Best model
}
\keyword{datasets}
