% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LEGIT.R
\name{example_3way}
\alias{example_3way}
\title{Simulated example of a 3 way interaction GxExz model}
\usage{
example_3way(N, sigma = 2.5, logit = FALSE, seed = NULL)
}
\arguments{
\item{N}{Sample size.}

\item{sigma}{Standard deviation of the gaussian noise (if \code{logit}=FALSE).}

\item{logit}{If TRUE, the outcome is transformed to binary with a logit link.}

\item{seed}{RNG seed.}
}
\value{
Returns a list containing, in the following order: data.frame with the observed outcome (with noise), the true outcome (without noise) and \eqn{z}, data.frame of the genetic variants (G), data.frame of the environments (E), vector of the true genetic coefficients, vector of the true environmental coefficients, vector of the true main model coefficients
}
\description{
Simulated example of a 3 way interaction GxExz model (where G and E are latent variables). 
\deqn{g_j \sim Binomial(n=1,p=.30)}
\deqn{j = 1, 2, 3, 4}
\deqn{e_l \sim Normal(\mu=0,\sigma=1.5)}
\deqn{l = 1, 2, 3}
\deqn{z \sim Normal(\mu=3,\sigma=1)}
\deqn{g = .2g_1 + .15g_2 - .3g_3 + .1g_4 + .05g_1g_3 + .2g_2g_3}
\deqn{e = -.45e_1 + .35e_2 + .2e_3}
\deqn{\mu = -2 + 2g + 3e + z + 5ge - 1.5ez + 2gz + 2gez}
\tabular{cc}{
\eqn{y \sim Normal(\mu=\mu,\sigma=\code{sigma})} if \code{logit}=FALSE \cr
\eqn{y \sim Binomial(n=1,p=logit(\mu))} if \code{logit}=TRUE
}
}
\examples{
example_3way(5,2.5,logit=FALSE)
example_3way(5,0,logit=TRUE)
}
