% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LEGIT.R
\docType{data}
\name{forward_step}
\alias{forward_step}
\title{Internal function that does the forward step for the stepwise function.}
\format{An object of class \code{function} of length 1.}
\usage{
forward_step
}
\arguments{
\item{empty_start_dataset}{If TRUE, the initial dataset is empty.}

\item{fit}{Current best fit.}

\item{...}{Same parameters as in the stepwise function.}
}
\value{
Returns fit, start_genes, start_env and genes_current, genes_toadd if search="genes" or env_current and env_toadd if search="env".
}
\description{
Internal function that does the forward step for the stepwise function.
}
\keyword{internal}
