% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LEGIT.R
\docType{data}
\name{forward_step_IM}
\alias{forward_step_IM}
\title{Internal function that does the forward step for the stepwise function.}
\format{An object of class \code{function} of length 1.}
\usage{
forward_step_IM
}
\arguments{
\item{empty_start_dataset}{If TRUE, the initial dataset is empty.}

\item{fit}{Current best fit.}

\item{...}{Same parameters as in the stepwise function.}
}
\value{
Returns fit, start_latent_var, latent_var_current and latent_var_toadd.
}
\description{
Internal function that does the forward step for the stepwise function.
}
\keyword{internal}
