% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LEGIT.R
\name{plot.elastic_net_var_select}
\alias{plot.elastic_net_var_select}
\title{Plot function for the output of elastic_net_var_select}
\usage{
\method{plot}{elastic_net_var_select}(x, lwd = 2, start = 1, ...)
}
\arguments{
\item{x}{An object of class "elastic_net_var_select", usually, a result of a call to elastic_net_var_select.}

\item{lwd}{Thickness of the lines (Default = 2)}

\item{start}{At which lambda to start (from large lambda to small lambda). If start is not 1, we remove some of the large lambda, this can make plot easier to visualize (Default = 1).}
}
\value{
Returns the plot of the coefficients of variables inside the latent variables with respect to the log(lambda).
}
\description{
Plot of the coefficients of variables inside the latent variables with respect to the log(lambda). This is your typical elastic-net plot.
}
\examples{
train = example_2way(500, 1, seed=777)
fit_best = IMLEGIT(train$data, list(G=train$G, E=train$E), y ~ G*E, 
list(train$coef_G, train$coef_E))
fit_default = IMLEGIT(train$data, list(G=train$G, E=train$E), y ~ G*E)
summary(fit_default)
summary(fit_best)
train = example_3way_3latent(500, 1, seed=777)
fit_best = IMLEGIT(train$data, train$latent_var, y ~ G*E*Z, 
list(train$coef_G, train$coef_E, train$coef_Z))
fit_default = IMLEGIT(train$data, train$latent_var, y ~ G*E*Z)
summary(fit_default)
summary(fit_best)
}
\references{
Alexia Jolicoeur-Martineau, Ashley Wazana, Eszter Szekely, Meir Steiner, Alison S. Fleming, James L. Kennedy, Michael J. Meaney, Celia M.T. Greenwood and the MAVAN team. \emph{Alternating optimization for GxE modelling with weighted genetic and environmental scores: examples from the MAVAN study} (2017). arXiv:1703.08111.
}
