% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LEGIT.R
\name{summary.IMLEGIT}
\alias{summary.IMLEGIT}
\title{Summarizing IMLEGIT fits}
\usage{
\method{summary}{IMLEGIT}(object, ...)
}
\arguments{
\item{object}{An object of class "IMLEGIT", usually, a result of a call to LEGIT.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Returns a list of objects of class "summary.glm" containing the summary of each parts (main and latent variables) of the model.
}
\description{
Shows the summary for all parts (main and latent variables) of the LEGIT model.
}
\examples{
	train = example_2way(250, 1, seed=777)
fit_default = IMLEGIT(train$data, list(G=train$G, E=train$E), y ~ G*E)
summary(fit_default)
}
