% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LEGIT.R
\name{backward_step}
\alias{backward_step}
\title{Internal function that does the backward step for the stepwise function.}
\usage{
backward_step(fit, data, formula, interactive_mode = FALSE,
  genes_current = NULL, env_current = NULL, genes_dropped = NULL,
  env_dropped = NULL, search = "genes", search_criterion = "AIC",
  p_threshold = 0.2, exclude_worse_AIC = TRUE, max_steps = 50,
  cv_iter = 5, cv_folds = 10, folds = NULL, classification = FALSE,
  start_genes = NULL, start_env = NULL, eps = 0.01, maxiter = 25,
  family = gaussian, seed = NULL, print = TRUE)
}
\arguments{
\item{fit}{Current best fit.}

\item{empty_start_dataset}{If TRUE, the initial dataset is empty.}

\item{...}{Same parameters as in the stepwise function.}
}
\value{
Returns fit, start_genes, start_env and genes_current, genes_toadd if search="genes" or env_current and env_toadd if search="env".
}
\description{
Internal function that does the backward step for the stepwise function.
}
\keyword{internal}
