% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LEGIT.R
\name{backward_step_IM}
\alias{backward_step_IM}
\title{Internal function that does the backward step for the stepwise IM function.}
\usage{
backward_step_IM(fit, data, formula, interactive_mode = FALSE,
  latent_var_current = NULL, latent_var_dropped = NULL, search = NULL,
  search_criterion = "AIC", p_threshold = 0.2, exclude_worse_AIC = TRUE,
  max_steps = 100, cv_iter = 5, cv_folds = 10, folds = NULL,
  Huber_p = 1, classification = FALSE,
  start_latent_var = start_latent_var, eps = 0.01, maxiter = 25,
  family = gaussian, seed = NULL, print = TRUE)
}
\arguments{
\item{fit}{Current best fit.}

\item{empty_start_dataset}{If TRUE, the initial dataset is empty.}

\item{...}{Same parameters as in the stepwise function.}
}
\value{
Returns fit, start_latent_var, latent_var_current and latent_var_dropped.
}
\description{
Internal function that does the backward step for the stepwise IM function.
}
\keyword{internal}
