% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LEGIT.R
\name{best_model}
\alias{best_model}
\title{Best model}
\usage{
best_model(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Best model
}
\description{
Best model
}
